fixup! fixup! HAI-753 Send täydennys attachments to Allu.
